share function update elems' layers
