+++ src/web/app/components/beambox/right-panel/SelLayerBlock.tsx
@@ -1,10 +1,8 @@
 import React, { useContext, memo } from 'react';
 
-import Alert from 'app/actions/alert-caller';
-import AlertConstants from 'app/constants/alert-constants';
 import i18n from 'helpers/i18n';
 import { CanvasContext } from 'app/contexts/CanvasContext';
-import { getObjectLayer } from 'helpers/layer/layer-helper';
+import { getObjectLayer, moveToOtherLayer } from 'helpers/layer/layer-helper';
 import { getSVGAsync } from 'helpers/svg-editor-helper';
 
 let svgCanvas;
@@ -31,31 +29,10 @@ function SelLayerBlock(): JSX.Element {
     options.push(<option value={layerName} key={i}>{layerName}</option>);
   }
 
-  const moveToOtherLayer = (e: React.ChangeEvent): void => {
+  const onChange = (e: React.ChangeEvent) => {
     const select = e.target as HTMLSelectElement;
     const destLayer = select.options[select.selectedIndex].value;
-    const confirmStr = LANG.notification.QmoveElemsToLayer.replace('%s', destLayer);
-    const moveToLayer = (ok) => {
-      if (!ok) {
-        return;
-      }
-      svgCanvas.moveSelectedToLayer(destLayer);
-      drawing.setCurrentLayer(destLayer);
-      setPromptMoveLayerOnce(true);
-    };
-
-    if (destLayer) {
-      if (promptMoveLayerOnce) {
-        moveToLayer(true);
-      } else {
-        Alert.popUp({
-          id: 'move layer',
-          buttonType: AlertConstants.YES_NO,
-          message: confirmStr,
-          onYes: moveToLayer,
-        });
-      }
-    }
+    moveToOtherLayer(destLayer, () => setPromptMoveLayerOnce(true), !promptMoveLayerOnce);
   };
 
   const currentLayerName = currentLayer.title;
@@ -66,7 +43,7 @@ function SelLayerBlock(): JSX.Element {
         value={currentLayerName}
         id="selLayerNames"
         title="Move selected elements to a different layer"
-        onChange={(e: React.ChangeEvent) => moveToOtherLayer(e)}
+        onChange={onChange}
         disabled={options.length < 2}
       >
         {options}
